StatusMessage is not always defined in the progress_event
With MemoryDB, `StatusMessage` is not always defined and this trigger an exception
within the existing exception handling block.

With this change we fallback on `OperationStatus` if `StatusMessage`
is missing. If none of them are available, we will get a `None` instead of
an exception.

According to the documentation the presence of these two keys is not guarantee.
They are both identified as optional (`Required: No`).

See: https://docs.aws.amazon.com/cloudcontrolapi/latest/APIReference/API_ProgressEvent.html